refactor: handle barcode correction logic at page level
